Apps/demerzel-agent/pyproject.toml
[project]
name = "demerzel-agent"
version = "1.0.0"
description = "Demerzel — AI Governance Agent (ACP protocol)"
requires-python = ">=3.12"
dependencies = [
    "acp-sdk>=0.2.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
demerzel = "src.server:main"
